+++ src/all/background_page/event/exportPasswordsIframeEvents.js
@@ -6,7 +6,7 @@
  * 	  will detect it and will display the iframe content.
  * 	- Close the iframe.
  *
- * @copyright (c) 2017 Passbolt SARL
+ * @copyright (c) 2017-2018 Passbolt SARL, 2019 Passbolt SA
  * @licence GNU Affero General Public License http://www.gnu.org/licenses/agpl-3.0.en.html
  */
 var Worker = require('../model/worker');
@@ -16,6 +16,7 @@ var listen = function (worker) {
 	/*
 	 * Open the export passwords dialog.
 	 * @listens passbolt.export-passwords.open-dialog
+   * @todo might be @deprecated
 	 */
 	worker.port.on('passbolt.export-passwords.open-dialog', function () {
 		Worker.get('App', worker.tab.id).port.emit('passbolt.export-passwords.open-dialog');